Unfortunatelly got rid of PPS stuff
CYPD3177 is not happy with PPS PDDs. Even if I craft a correct RDO for a
PPS PDO (and the source responds with ACCEPT & then READY), CYPD3177
freaks out and resets immediatelly after receiving the accept and ready
message for a PPS PDO. So I believe it would be hard to hack around and
implement PPS with it, sadly.